requirements.txt
annotated-types==0.7.0
attrs==25.4.0
bandit==1.8.6
blinker==1.9.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fastjsonschema==2.21.2
Flask==3.1.2
fonttools==4.60.1
gunicorn==23.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mypy_extensions==1.1.0
narwhals==2.9.0
nbformat==5.10.4
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandera==0.26.1
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rich==14.2.0
rpds-py==0.28.0
ruff==0.14.2
six==1.17.0
stevedore==5.5.0
traitlets==5.14.3
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
